refactor: move `storedWalletState`, `pollOffer` to `lib/wallet.js`
