+++ packages/agoric-cli/src/lib/wallet.js
@@ -3,11 +3,10 @@
 
 import { iterateReverse } from '@agoric/casting';
 import { makeWalletStateCoalescer } from '@agoric/smart-wallet/src/utils.js';
-import { execSwingsetTransaction, pollTx } from './chain.js';
-import { boardSlottingMarshaller } from './rpc.js';
+import { execSwingsetTransaction, pollBlocks, pollTx } from './chain.js';
+import { boardSlottingMarshaller, makeRpcUtils } from './rpc.js';
 
 /** @typedef {import('@agoric/smart-wallet/src/smartWallet.js').CurrentWalletRecord} CurrentWalletRecord  */
-/** @typedef {import('@agoric/smart-wallet/src/offers.js').OfferSpec} OfferSpec */
 /** @typedef {import('@agoric/vats/tools/board-utils.js').AgoricNamesRemotes} AgoricNamesRemotes  */
 
 const marshaller = boardSlottingMarshaller();
@@ -158,4 +157,69 @@ export const findContinuingIds = (current, agoricNames) => {
     }
   }
   return found;
+};
+
+export const makeWalletUtils = async (
+  { fetch, execFileSync, delay },
+  networkConfig,
+) => {
+  const { agoricNames, fromBoard, readLatestHead, vstorage } =
+    await makeRpcUtils({ fetch }, networkConfig);
+  /**
+   * @param {string} from
+   * @param {number|string} [minHeight]
+   */
+  const storedWalletState = async (from, minHeight = undefined) => {
+    const m = boardSlottingMarshaller(fromBoard.convertSlotToVal);
+
+    const history = await vstorage.readFully(
+      `published.wallet.${from}`,
+      minHeight,
+    );
+
+    /** @type {{ Invitation: Brand<'set'> }} */
+    // @ts-expect-error XXX how to narrow AssetKind to set?
+    const { Invitation } = agoricNames.brand;
+    const coalescer = makeWalletStateCoalescer(Invitation);
+    // update with oldest first
+    for (const txt of history.reverse()) {
+      const { body, slots } = JSON.parse(txt);
+      const record = m.unserialize({ body, slots });
+      coalescer.update(record);
+    }
+    const coalesced = coalescer.state;
+    harden(coalesced);
+    return coalesced;
+  };
+
+  /**
+   * Get OfferStatus by id, polling until available.
+   *
+   * @param {string} from
+   * @param {string|number} id
+   * @param {number|string} minHeight
+   */
+  const pollOffer = async (from, id, minHeight) => {
+    const lookup = async () => {
+      // eslint-disable-next-line @jessie.js/no-nested-await, no-await-in-loop
+      const { offerStatuses } = await storedWalletState(from, minHeight);
+      const offerStatus = [...offerStatuses.values()].find(s => s.id === id);
+      if (!offerStatus) throw Error('retry');
+      harden(offerStatus);
+      return offerStatus;
+    };
+    const retryMessage = 'offer not in wallet at block';
+    const opts = { ...networkConfig, execFileSync, delay, retryMessage };
+    return pollBlocks(opts)(lookup);
+  };
+
+  return {
+    networkConfig,
+    agoricNames,
+    fromBoard,
+    vstorage,
+    readLatestHead,
+    storedWalletState,
+    pollOffer,
+  };
 };
